+++ ppapi/proxy/url_loader_resource.cc
@@ -238,7 +238,7 @@ void URLLoaderResource::OnPluginMsgSendData(
     const IPC::Message& message) {
   base::PickleIterator iter(message);
   const char* data;
-  int data_length;
+  size_t data_length;
   if (!iter.ReadData(&data, &data_length)) {
     NOTREACHED() << "Expecting data";
     return;
